a2abackend/pyproject.toml
[project]
name = "version_3_multi_agent"
version = "0.1.0"
description = "version_3_multi_agent"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "asyncclick>=8.1.8",
    "click>=8.1.8",
    "fastapi>=0.115.12",
    "google-adk>=1.0.0",
    "google-genai>=1.11.0",
    "httpx>=0.28.1",
    "httpx-sse>=0.4.0",
    "psycopg2-binary>=2.9.9",
    "pydantic>=2.11.4",
    "python-dotenv>=1.0.1",
    "starlette>=0.46.2",
    "uvicorn>=0.34.2",
           "hedera-sdk-py>=2.50.0",
    "web3>=6.0.0",
    "eth-account>=0.8.0",
    "requests>=2.31.0",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["agents*", "app*", "client*", "models*", "server*", "utilities*"]
